requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
pendulum==3.1.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
